pyproject.toml
[project]
name = "beeek18-translate-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "aiogram>=3.26.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.4",
    "watchfiles>=1.1.1",
]
